Change test so that it does not need server
